requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
requests==2.31.0
pandas==2.0.2
sqlalchemy==2.0.21
numpy==1.24.4
beautifulsoup4==4.12.2
gunicorn==21.2.0
python-dotenv==1.0.0
supabase==2.1.0
redis==5.0.8
nltk==3.8.1
flask-cors==3.0.10